requirements.txt
python-telegram-bot
httpx
python-dotenv
jinja2 
weasyprint